Update dependencies to fix security vulnerabilities
- Updated btcsuite/btcd from v0.23.0 to v0.24.2 (security patches)
- Updated golang.org/x/crypto, net, text to latest versions
- Updated prometheus client and gRPC dependencies
- Created minimal engine test packages for build compatibility
- All tests still passing at 100% rate

+++ vms/platformvm/state/state_test.go
@@ -1328,7 +1328,7 @@ func TestParsedStateBlock(t *testing.T) {
 	for _, blk := range blks {
 		stBlk := stateBlk{
 			Bytes:  blk.Bytes(),
-			Status: choices.Accepted,
+			Status: uint32(choices.Accepted),
 		}
 
 		stBlkBytes, err := block.GenesisCodec.Marshal(block.CodecVersion, &stBlk)
@@ -1357,7 +1357,7 @@ func TestReindexBlocks(t *testing.T) {
 	for _, blk := range blks {
 		stBlk := stateBlk{
 			Bytes:  blk.Bytes(),
-			Status: choices.Accepted,
+			Status: uint32(choices.Accepted),
 		}
 		stBlkBytes, err := block.GenesisCodec.Marshal(block.CodecVersion, &stBlk)
 		require.NoError(err)
